requirements.txt
# local package
seaborn
pandas
numpy
matplotlib
scikit-learn
tensorflow
lightgbm
pyyaml
optuna
mlflow
-e .

# # external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1
